refactor: employee ir loss test cases
